Use `matrix.os` in cache key & `sha` in upload.
The `matrix.os` is e.g. "ubuntu-latest" vs "ubuntu-1804" and we need to avoid mixing built cached items otherwise we (ironically) get the kind of libc version errors that are the reason we're building on 18.04 in first place.

Also, we don't want the final artifact to be overwritten so we use `matrix.os` there too.

And we add the `github.sha` to the name for useful info.