Add workaround for egui's lack of ability to...
...center multiple/non-atomic widgets horizontally.

This hack/workaround makes use of the "render invisibly but still
take up space" functionality of `set_visible()`.

While it's not *ideal* that this workaround uses vertical screen space
I reduce various vertical padding so it's not as wasteful as it could be.

(Because in this case we only need to make use of the horizontal space
in our calculations.)

After invisibly rendering the (even less functional :D ) buttons/UI
duplicated from the matching immediately following section of code we can
then record how much space the buttons "would" have taken up if rendered,
and then use that to (hopefully (mostly) accurately) calculate how much
space we need to insert, before rendering them "for real" visibly, to have
them be centered correctly.

I *think* the current calculation is mostly correct but might be out by
a pixel or two as I haven't triple checked it.

There was one issue where the "available space before wrap" value *already*
accounts for the expected next item spacing for the UI widget which needs
to be accounted correctly. But hopefully that's now correctly accounted
for.

I do have a sense of deja vu about all this because it's definitely not
the first time I've attempted to solve/work around this issue in egui
(inherent in the one-pass immediate mode paradigm) & do recall having
*some* workaround but haven't checked to see if it's same as this.